Fix bug with parameter schema stuff
